ref: moved functions to appropriated module
